Run version check on all subdirectories of METADATA
This PR addresses an edge case observed in #2095 in which a package was
registered and the user attempted to register versions manually but
placed sha1 under "METADATA/$pkg/versions" instead of
"METADATA/$pkg/versions/v0.01".

The version check test introduced by #2040 is now placed in a new loop
that runs for each subdirectory of METADATA, regardless of whether each
subdirectory corresponds to a package entry recognized by Pkg.Read.
Doing so allows the tests to detect package entries that are
misconfigured and hence not recognized as such by the package manager.

Additionally, joinpath is used consistently instead of manually splicing
strings together.

Furthermore, as PEGParsers is the only package left without a .jl
extension in #2057, the .jl extension check is now upgraded to an
assertion error except for PEGParsers, which is grandfathered in.