An upcoming change in package PDMats to support floating point types other than Float64 means that some typealiases in the Distributions package need updating. This fix to METADATA fixes the upper release of PDMats requirements for the Distributions and the ConjugatePriors package to the upcoming version number for PDMats (0.4).
The relevant issue in PDMats i::
https://github.com/JuliaStats/PDMats.jl/pull/37

I checked the dependencies of the PDMats package with ``git grep PDMats`` in
METADATA, and this are the results for the other packages which declare a
dependence:

Packages Distributions and ConjugatePriors are fixed with this submit.
Packages GaussianMixtures, Mamba and MixedModels all pass their tests.
Packages Brownian and DirichletProcessMixtures are not in development and
appear broken for other reasons.