Maximum versions on DiffEqBase for StochasticDiffEq
There will be a tag in DiffEqBase and a corresponding update in StochasticDiffEq which changes the noise interface to allow for in-place RNGs. No previous version will be compatible with the defaults in the new version, so this will block the release until a compatible version of StochasticDiffEq.jl is available.