Update URL
Hi, I've been working on bringing FactCheck.jl up-to-date to be compatible with Julia v1.0 and Julia v1.1. I've edited the source to use the updated syntax and have done my best to get runtests.jl to run and succeed in the travis.cl config file. As far as I can tell, my package is using the new Pkg format correctly in terms of the Project.toml and Manifest files. 

If I'm missing anything or if the updated package is not up to snuff please let me know and I'll try to correct any issues. 

Thanks you